🚚 Code splitting for rarity requests
